+++ src/AutomaticRoomIdentification/Eigen/src/Core/util/Memory.h
@@ -8,7 +8,7 @@
 // Copyright (C) 2010 Thomas Capricelli <orzel@freehackers.org>
 // Copyright (C) 2013 Pavel Holoborodko <pavel@holoborodko.com>
 //
-// This Source Code Form is subject to the terms of the Mozilla
+// This SourceMode Code Form is subject to the terms of the Mozilla
 // Public License v. 2.0. If a copy of the MPL was not distributed
 // with this file, You can obtain one at http://mozilla.org/MPL/2.0/.